+++ src/apps/es/items--search--empty/index.tsx
@@ -1,0 +1,46 @@
+/* ======================================================================== *
+ * Copyright 2024 HCL America Inc.                                          *
+ * Licensed under the Apache License, Version 2.0 (the "License");          *
+ * you may not use this file except in compliance with the License.         *
+ * You may obtain a copy of the License at                                  *
+ *                                                                          *
+ * http://www.apache.org/licenses/LICENSE-2.0                               *
+ *                                                                          *
+ * Unless required by applicable law or agreed to in writing, software      *
+ * distributed under the License is distributed on an "AS IS" BASIS,        *
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. *
+ * See the License for the specific language governing permissions and      *
+ * limitations under the License.                                           *
+ * ======================================================================== */
+
+import { createSvgIcon, IIconAttrs, IIconContent } from '../../../utils';
+
+const attrs: IIconAttrs = {
+  xmlns: 'http://www.w3.org/2000/svg',
+  viewBox: '0 0 32 32',
+  width: 32,
+  height: 32,
+};
+
+const content: IIconContent[] = [
+  {
+    elem: 'path',
+    attrs: {
+      d:
+        `M13.3597 0.720002H11.7197V2.36H10.9997V0.36C10.9997 0.161 11.1607 0 11.3597 0H13.3597V0.720002ZM16.0067 0H19.5367V0.720002H16.0067V0ZM25.7117 
+        0H22.1827V0.720002H25.7117V0ZM28.3597 0H30.3597C30.5587 0 30.7197 0.161 30.7207 0.36V2.36H29.9997V0.720002H28.3597V0ZM30.7207 12.948H29.9997V17.183H30.7207V12.948ZM29.9997 
+        5.53594H30.7207V9.77095H29.9997V5.53594ZM29.9997 22.0001H28.3597V22.7211H30.3597C30.5587 22.7211 30.7197 22.5601 30.7197 22.3611V20.3611H29.9997V22.0001ZM23.4607 
+        22H26.2607V22.721H23.4607V22ZM11.7197 8.29958V6.36006H10.9997V8.14393C10.4616 8.04947 9.90745 8.00035 9.34147 8.00035C4.17643 8.00035 0 12.1961 0 17.3849C0 22.5737 4.17643 
+        26.7694 9.34147 26.7694C11.7526 26.7694 14.0737 25.8633 15.723 24.2291L22.4633 31.0004L23 30.4611L16.2494 23.6794C17.7208 22.0531 18.683 19.8643 18.683 17.3849C18.683 12.9307 
+        15.7286 9.33251 11.7197 8.29958ZM21.3597 22.7201H18.36V22.0001H21.3597V22.7201ZM9.34147 8.76297C4.59569 8.76297 0.759122 12.6173 0.759122 17.3849C0.759122 22.1525 4.59569 
+        26.0068 9.34147 26.0068C11.6704 26.0068 13.8867 25.0908 15.3907 23.4793L15.399 23.4703C16.9153 21.947 17.9239 19.8176 17.9239 17.3849C17.9239 12.5143 14.0902 8.76297 9.34147 
+        8.76297ZM12.2357 13.8308L9.34147 16.7383L6.44624 13.8308L5.90958 14.371L8.80375 17.2785L5.90958 20.186L6.44624 20.7261L9.34147 17.8186L12.2357 20.7261L12.7734 20.186L9.87915 
+        17.2785L12.7734 14.371L12.2357 13.8308Z`,
+      fill: 'currentColor',
+      fillRule: 'evenodd',
+      clipRule: 'evenodd',
+    },
+  },
+];
+
+export default createSvgIcon('items--search--empty', 32, content, attrs);